requirements.txt
aiohttp==3.7.3
aiohttp-jinja2==1.4.2
aiohttp-session==2.9.0
cryptography==3.4.7
discord.py==1.6.0
ffmpeg==1.4
Flask==0.10.1
Flask-Runner==2.1.1
Flask-Script==0.6.7
Jinja2==2.11.3
mutagen==1.45.1
oauthlib==3.1.0
PyNaCl==1.4.0
python-dotenv==0.15.0
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
urllib3==1.26.3
google-api-core==1.31.1
google-auth==1.34.0
google-cloud==0.34.0
google-cloud-vision==2.4.2
googleapis-common-protos==1.53.0
tinydb==4.5.1vi 